# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DraftPatch": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "cc": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cc"
          },
          "current_draft_id": {
            "format": "uuid",
            "title": "Current Draft Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "subject": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "thread_id": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "to": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "To"
          }
        },
        "required": [
          "expected_version",
          "current_draft_id"
        ],
        "title": "DraftPatch",
        "type": "object"
      },
...
      "EmailState": {
        "enum": [
          "INGESTED",
          "DRAFTING",
          "DRAFT_RETRY_WAIT",
          "AWAITING_REVIEW",
          "APPROVED",
          "SEND_PENDING",
          "SENDING",
          "SENT",
          "REJECTED",
          "SEND_RETRY_WAIT",
          "DELIVERY_UNKNOWN",
          "FAILED_TERMINAL"
        ],
        "title": "EmailState",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewActionPayload": {
        "additionalProperties": false,
        "properties": {
          "current_draft_id": {
            "format": "uuid",
            "title": "Current Draft Id",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version",
          "current_draft_id"
        ],
        "title": "ReviewActionPayload",
        "type": "object"
      },
...
      "StaffEmailQueueItem": {
        "additionalProperties": false,
        "description": "Safe staff-facing projection of an email work item.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmailPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "sender": {
            "title": "Sender",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/EmailState"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "state",
          "version",
          "sender",
          "subject",
          "received_at",
          "category",
          "priority"
        ],
        "title": "StaffEmailQueueItem",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "PublicChatCredential": {
        "bearerFormat": "session credential",
        "description": "Opaque credential issued for one public chat session.",
        "scheme": "bearer",
        "type": "http"
      },
      "StaffSession": {
        "description": "Issued after approved staff OIDC admission; state-changing routes also require CSRF.",
        "in": "cookie",
        "name": "staff_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Enterprise AI Agent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/staff/email": {
      "get": {
        "operationId": "list_email_queue_api_v1_staff_email_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/EmailState"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StaffEmailQueueItem"
                  },
                  "title": "Response List Email Queue Api V1 Staff Email Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "List Email Queue",
        "tags": [
          "email-review"
        ],
        "x-security-classification": "staff"
      }
    },
...
    "/api/v1/staff/email/{work_item_id}": {
      "get": {
        "operationId": "read_email_detail_api_v1_staff_email__work_item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffEmailDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Read Email Detail",
        "tags": [
          "email-review"
        ],
        "x-security-classification": "staff"
      }
    },
    "/api/v1/staff/email/{work_item_id}/approve": {
      "post": {
        "operationId": "approve_api_v1_staff_email__work_item_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Api V1 Staff Email  Work Item Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Approve",
        "tags": [
          "email-review"
        ],
        "x-security-classification": "staff"
      }
    },
    "/api/v1/staff/email/{work_item_id}/draft": {
      "patch": {
        "operationId": "edit_draft_api_v1_staff_email__work_item_id__draft_patch",
        "parameters": [
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Edit Draft Api V1 Staff Email  Work Item Id  Draft Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Edit Draft",
        "tags": [
          "email-review"
        ],
        "x-security-classification": "staff"
      }
    },
    "/api/v1/staff/email/{work_item_id}/regenerate": {
      "post": {
        "operationId": "regenerate_api_v1_staff_email__work_item_id__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegeneratePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Regenerate Api V1 Staff Email  Work Item Id  Regenerate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Regenerate",
        "tags": [
          "email-review"
        ],
        "x-security-classification": "staff"
      }
    },
    "/api/v1/staff/email/{work_item_id}/reject": {
      "post": {
        "operationId": "reject_api_v1_staff_email__work_item_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          },
          {
            "in": "cookie",
            "name": "staff_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Staff Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewActionPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reject Api V1 Staff Email  Work Item Id  Reject Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "StaffSession": []
          }
        ],
        "summary": "Reject",
        "tags": [
          "email-review"
        ],
        "x-security-classification": "staff"
      }
    },
...
  }
}